+++ uplink/src/base/middleware/logcat.rs
